# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TechLane API
  version: 1.0.0
  description: Shared REST API for TechLane ops web, Android, and portals.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      security: []
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
  /auth/refresh:
    post:
      security: []
      summary: Refresh access token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Tokens rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
...
  /users:
    get:
      summary: List staff
      operationId: listUsers
      parameters:
        - name: role
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Staff list
    post:
      summary: Create staff
      operationId: createUser
      responses:
        "201":
          description: Created
  /users/{id}:
    get:
      summary: Get staff member
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Staff detail
    patch:
      summary: Update staff roles, branches, status
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated
  /users/{id}/commission:
    put:
      summary: Set optional technician commission
      operationId: setUserCommission
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated
...
  /commissions:
    get:
      summary: List commission entries
      operationId: listCommissions
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: user_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Commission ledger
  /commissions/{id}/approve:
    post:
      summary: Approve pending commission
      operationId: approveCommission
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Approved
  /commissions/{id}/mark-paid:
    post:
      summary: Mark commission paid
      operationId: markCommissionPaid
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Paid
...
  /devices/{id}/revoke:
    post:
      summary: Revoke a registered device
      operationId: revokeDevice
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Revoked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
...